Restructure to have game.vue page
