+++ wordle-web/app.vue
@@ -2,7 +2,7 @@
   <NuxtLayout>
     <v-app>
       <v-btn @click="router.push('/')">Home</v-btn>
-      <v-btn @click="router.push('/test')">Test</v-btn>
+      <v-btn @click="router.push('/game')">WordGuess</v-btn>
       <NuxtPage />
     </v-app>
   </NuxtLayout>
