--- wordle-web/pages/test.vue
@@ -1,5 +1,0 @@
-<template>
-    <v-card>
-        <v-card-title>Hello Welcome to Test</v-card-title>
-    </v-card>
-</template>
